Logs

+++ src/endpoints/test-utils/helpers/mongo.ts
@@ -3,23 +3,17 @@ import {map} from 'lodash';
 import {Connection} from 'mongoose';
 import {waitOnPromisesAndLogErrors} from '../../../utilities/promiseFns';
 
-export async function dropMongoConnection(
-  connection: Connection,
-  dropDb = false
-) {
-  // console.log(
-  //   `-- Mongo - dropping mongo collections in db ${connection.db.databaseName} --`
-  // );
+export async function dropMongoConnection(connection?: Connection | null) {
+  if (!connection) {
+    return;
+  }
+
   const collections = await connection.db.collections();
   const promises = map(collections, collection => {
     return collection.drop();
   });
 
   await waitOnPromisesAndLogErrors(promises);
-  // if (dropDb) {
-  //   await connection.db.dropDatabase();
-  // }
-
   await connection.db.dropDatabase();
   await connection.close();
 }
